Split out errors into error.rs
