+++ src/lib.rs
@@ -1,7 +1,10 @@
 // #![deny(missing_docs)]
 
 pub mod brb_membership;
-pub use crate::brb_membership::{Ballot, Error, Generation, Reconfig, State, Vote, VoteMsg};
+pub use crate::brb_membership::{Ballot, Generation, Reconfig, State, Vote, VoteMsg};
+
+pub mod error;
+pub use crate::error::Error;
 
 pub mod actor;
 pub use actor::{Actor, Sig, SigningActor};
